pyproject.toml
[tool.poetry]
name = "pep-opex"
version = "0.1.0"
description = ""
authors = ["spogins <spogins@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
wagtail = "^5.2.2"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
django-vite = "^3.0.2"
wagtail-modeltranslation = "^0.14.1"
wagtailvideos = "^5.2.1"
wagtail-seo = "^2.5.0"
wagtail-robots = "^0.4.0"
wagtail-modeladmin = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
